ADRIA API

Metrics

Performance indicators

ADRIA.performance.RMSE — Method

Root Mean Square Error

ADRIA.performance.environmental_diversity — Method

environmental_diversity(ms, inputs_i)

Obtain an indication of environmental factor diversity for a scenario set. Higher values indicate a greater of mix of environmental conditions were experienced between scenarios.

This is referred to as $E$.

Arguments

ms : model spec

inputs_i : inputs used for scenarios of interest

ADRIA.performance.gmd — Method

gmd(vals::AbstractVector{<:Real})::Float64
gmd(vals::AbstractMatrix{<:Real})

Gini's Mean Difference.

The absolute mean of all pairwise distances between elements in a given set.

References

La Haye, R., & Zizler, P. (2019). The Gini mean difference and variance. METRON, 77(1), 43-52. https://doi.org/10.1007/s40300-019-00149-2

Yitzhaki, S. (2003). Gini's Mean difference: A superior measure of variability for non-normal distributions. Metron - International Journal of Statistics, LXI(2), 285-316. https://ideas.repec.org/a/mtn/ancoec/030208.html

Kashif, M., Aslam, M., Al-Marshadi, A. H., & Jun, C.-H. (2016). Capability Indices for Non-Normal Distribution Using Gini's Mean Difference as Measure of Variability. IEEE Access, 4, 7322-7330. https://doi.org/10.1109/ACCESS.2016.2620241

ADRIA.performance.intervention_diversity — Method

intervention_diversity(ms, inputs_i)

Obtain an indication of intervention diversity for a scenario. Higher values indicate a greater of mix of interventions options were applied.

This is referred to as $D$.

Arguments

ms : model spec

inputs_i : inputs used for scenarios of interest

ADRIA.performance.intervention_effort — Method

intervention_effort(ms, inputs_i)

Obtain an indication of intervention effort for each scenario and intervention type. This is referred to as $F$.

Arguments

ms : model spec

inputs_i : inputs used for scenarios of interest

Returns

Matrix of s * 6, where s is the number of scenarios and columns are: seed_TA, seed_CA, fogging, SRM, seed_years, shade_years

ADRIA.performance.normalize — Method

normalize(vals::AbstractArray{<:Real})

Normalize values using feature scaling such that values are bound between 0 and 1, where 1 is equivalent to the maximum value found.

ADRIA.performance.probability — Method

probability(vals::AbstractArray{<:Real})

Calculate probability of individual trajectories, given a scenario ensemble $S$.

ADRIA.performance.temporal_variability — Method

temporal_variability(x::AbstractVector{<:Real})
temporal_variability(x::AbstractArray{<:Real, 2})
temporal_variability(x::AbstractArray{<:Real}, func_or_data...)

The V meta-metric.

As a meta-metric, it can be applied to any combination of metrics (including itself), assuming $x$ is bound between 0 and 1. If this is not the case, consider normalizing values first.

Examples

# Apply V to a time series
julia> temporal_variability(rand(50))

# Apply V to an ensemble of of time series
julia> x = rand(50, 200)
julia> temporal_variability(x)

# Create and apply a modified V metric to an ensemble of of time series.
# Where the argument is an array and not a function, the data is used directly
# and so it is assumed all matrices are of the same size and shape.
julia> temporal_variability(x, temporal_variabilty, temporal_variability(P(x)))
julia> temporal_variability(x, temporal_variabilty, P(x), D(x), E(x))

Sensitivity

ADRIA.sensitivity.ks_statistic — Method

ks_statistic(ks)

Calculate the Kolmogorov-Smirnov test statistic.

ADRIA.sensitivity.pawn — Method

Calculates the PAWN sensitivity index.

Arguments

X : Model inputs

Y : Outputs

S : Number of slides (default: 10)

Returns

NamedDimsArray, of min, mean, median, max, std, and cv summary statistics.

References

Pianosi, F., Wagener, T., 2018. Distribution-based sensitivity analysis from a generic input-output sample. Environmental Modelling & Software 108, 197-207. https://doi.org/10.1016/j.envsoft.2018.07.019

Baroni, G., Francke, T., 2020. GSA-cvd Combining variance- and distribution-based global sensitivity analysis https://github.com/baronig/GSA-cvd

ADRIA.sensitivity.relative_importance — Method

relative_importance(x::AbstractVector{<:Real})
relative_importance(x::AbstractArray)

Normalize metrics such that the values ∈ [0, 1].

Note: If a matrix is passed in, this calculates the relative importance for each column.

Arguments

x : metric values

Returns

Normalized values of same shape as x.

General API

ADRIA.bleaching_mortality! — Method

bleaching_mortality!(Y::Matrix{Float64}, tstep::Int64, depth::Vector{Float64},
    s::Vector{Float64}, dhw::Float64, a_adapt::Float64, n_adapt::Float64)

Calculates bleaching mortality taking into account depth and bleaching sensitivity of corals. Model is adapted from Bozec et al., [2], itself based on data from Hughes et al., 3 and Baird et al., 1.

Arguments

Y : Matrix to save results into

tstep : Current time step

depth : Mean site depth (m) for each site

s : Bleaching sensitivity of corals (relative values) for each taxa/size class

dhw : Degree Heating Week experienced at site

a_adapt : Level of assisted adaptation (DHW reduction)

n_adapt : Level of natural adaptation (DHW reduction linearly scaled over time)

Returns

Nothing

References

Baird, A., Madin, J., Álvarez-Noriega, M., Fontoura, L., Kerry, J., Kuo, C., Precoda, K., Torres-Pulliza, D., Woods, R., Zawada, K., & Hughes, T. (2018). A decline in bleaching suggests that depth can provide a refuge from global warming in most coral taxa. Marine Ecology Progress Series, 603, 257-264. https://doi.org/10.3354/meps12732

Bozec, Y.-M., Hock, K., Mason, R. A. B., Baird, M. E., Castro-Sanguino, C., Condie, S. A., Puotinen, M., Thompson, A., & Mumby, P. J. (2022). Cumulative impacts across Australia's Great Barrier Reef: A mechanistic evaluation. Ecological Monographs, 92(1), e01494. https://doi.org/10.1002/ecm.1494

Hughes, T. P., Kerry, J. T., Baird, A. H., Connolly, S. R., Dietzel, A., Eakin, C. M., Heron, S. F., Hoey, A. S., Hoogenboom, M. O., Liu, G., McWilliam, M. J., Pears, R. J., Pratchett, M. S., Skirving, W. J., Stella, J. S., & Torda, G. (2018). Global warming transforms coral reef assemblages. Nature, 556(7702), 492-496. https://doi.org/10.1038/s41586-018-0041-2

ADRIA.coral_spec — Method

coral_spec()

Template for coral parameter values for ADRIA. Includes "vital" bio/ecological parameters, to be filled with sampled or user-specified values.

Any parameter added to the params DataFrame defined here will automatically be made available to the ADRIA model.

Notes: Values for the historical, temporal patterns of degree heating weeks between bleaching years come from [1].

Returns

params : NamedTuple[taxanames, paramnames, params], taxa names, parameter names, and parameter values for each coral taxa, group and size class

References

Lough, J. M., Anderson, K. D., & Hughes, T. P. (2018). Increasing thermal stress for tropical coral reefs: 1871-2017. Scientific Reports, 8(1), 6079. https://doi.org/10.1038/s41598-018-24530-9

Hall, V.R. & Hughes, T.P. 1996. Reproductive strategies of modular organisms: comparative studies of reef-building corals. Ecology, 77: 950 - 963. https://dx.doi.org/10.2307/2265514

Bozec, Y.-M., Rowell, D., Harrison, L., Gaskell, J., Hock, K., Callaghan, D., Gorton, R., Kovacs, E. M., Lyons, M., Mumby, P., & Roelfsema, C. (2021). Baseline mapping to support reef restoration and resilience-based management in the Whitsundays. https://doi.org/10.13140/RG.2.2.26976.20482

Bozec, Y.-M., Hock, K., Mason, R. A. B., Baird, M. E., Castro-Sanguino, C., Condie, S. A., Puotinen, M., Thompson, A., & Mumby, P. J. (2022). Cumulative impacts across Australia's Great Barrier Reef: A mechanistic evaluation. Ecological Monographs, 92(1), e01494. https://doi.org/10.1002/ecm.1494

ADRIA.create_coral_struct — Function

create_coral_struct(bounds=(0.9, 1.1))

Generates Coral struct using the default parameter spec.

Example

# Define coral struct with auto-generated parameter ranges
# (default in ADRIA is ± 10%, triangular distribution with peak at 0.5)
create_coral_struct()
coral = Coral()

# Recreate coral spec ± 50% from nominal values
create_coral_struct((0.5, 1.5))
coral = Coral()

ADRIA.env_stats — Method

env_stats(rs::ResultSet, s_name::String, rcp::String)
env_stats(rs::ResultSet, s_name::String, rcp::String, member::Int)
env_stats(rs::ResultSet, s_name::String, stat::String, rcp::String, member::Int)

Extract statistics for a given environmental layer ("DHW" or "wave")

ADRIA.fecundity_scope! — Method

fecundity_scope!(fec_groups::Array{Float64, 2}, fec_all::Array{Float64, 2}, fec_params::Array{Float64},
                 Y_pstep::Array{Float64, 2}, k_area::Array{Float64})::Nothing

The scope that different coral groups and size classes have for producing larvae without consideration of environment.

Coral fecundity per coral area of the different size classes. When multiplied by the relative cover of each size class within taxa, this produces an estimate of the relative fecundity of each coral group and size. Total relative fecundity of a group is then calculated as the sum of fecundities across size classes.

Arguments

fecgroups : Matrix[nclasses, n_sites], memory cache to place results into

fecall : Matrix[ntaxa, n_sites], temporary cache to place intermediate fecundity values into

fec_params : Vector, coral fecundity parameters (in per m²) for each species/size class

Ypstep : Matrix[ntaxa, n_sites], of coral cover values for the previous time step

sitearea : Vector[nsites], total site area in m²

Returns

Matrix[nclasses, nsites] : fecundity per m² of coral

ADRIA.growthODE — Method

growthODE(du, X, p, _)

Base coral growth function.

ADRIA.run_scenario — Method

run_scenario(idx, param_set::Union{AbstractVector, DataFrameRow}, domain::Domain, data_store::NamedTuple, cache::NamedTuple)::NamedTuple
run_scenario(idx, param_set::Union{AbstractVector, DataFrameRow}, domain::Domain, data_store::NamedTuple)::NamedTuple
run_scenario(param_set::Union{AbstractVector, DataFrameRow}, domain::Domain, cache::NamedTuple)::NamedTuple
run_scenario(param_set::NamedTuple, domain::Domain)::NamedTuple

Run individual scenarios for a given domain, saving results to a Zarr data store. Results are stored in Zarr format at a pre-configured location. Sets up a new cache if not provided.

Notes

Logs of site ranks only store the mean site rankings over all environmental scenarios. This is to reduce the volume of data stored.

ADRIA.select — Method

select(r::ResultSet, op::String)

Hacky scenario filtering - to be replaced with more robust approach.

Only supports filtering by single attribute. Should be expanded to support filtering metric results too.

Examples

select(result, "guided .> 0.0")

# Above expands to:
# result.inputs.guided .> 0.0

ADRIA.site_area — Method

site_area(domain::Domain)::Vector{Float64}

Get site area for the given domain.

ADRIA.site_k_area — Method

site_k_area(domain::Domain)::Vector{Float64}

Get maximum coral cover area for the given domain in absolute area.

ADRIA.timesteps — Method

timesteps(rs::ResultSet)

Retrieve the time steps represented in the result set.

Arguments

rs : ResultSet

ADRIA.timesteps — Method

Extract the time steps represented in the data package.

ADRIA.Domain — Type

Domain{M,I,D,S,V,T,X}

Core ADRIA domain. Represents study area.

ADRIA.Domain — Method

Barrier function to create Domain struct without specifying Intervention/Criteria/Coral/SimConstant parameters.

ADRIA.Domain — Method

Domain(name::String, rcp::String, timeframe::Vector, site_data_fn::String, site_id_col::String, unique_site_id_col::String, init_coral_fn::String,
       conn_path::String, dhw_fn::String, wave_fn::String)::Domain

Convenience constructor for Domain.

Arguments

name : Name of domain

dpkg_path : location of data package

rcp : RCP scenario represented

timeframe : Time steps represented

sitedatafn : File name of spatial data used

siteidcol : Column holding name of reef the site is associated with (non-unique)

uniquesiteid_col : Column holding unique site names/ids

initcoralfn : Name of file holding initial coral cover values

conn_path : Path to directory holding connectivity data

dhw_fn : Filename of DHW data cube in use

wave_fn : Filename of wave data cube

ADRIA.SimConstants — Type

SimConstants

Struct of simulation constants for ADRIA

References

Lough, J. M., Anderson, K. D., & Hughes, T. P. (2018). Increasing thermal stress for tropical coral reefs: 1871-2017. Scientific Reports, 8(1), 6079. https://doi.org/10.1038/s41598-018-24530-9

Hughes, T. P., Kerry, J. T., Baird, A. H., Connolly, S. R., Dietzel, A., Eakin, C. M., Heron, S. F., Hoey, A. S., Hoogenboom, M. O., Liu, G., McWilliam, M. J., Pears, R. J., Pratchett, M. S., Skirving, W. J., Stella, J. S., & Torda, G. (2018). Global warming transforms coral reef assemblages. Nature, 556(7702), 492-496. https://doi.org/10.1038/s41586-018-0041-2

Bozec, Y.-M., Rowell, D., Harrison, L., Gaskell, J., Hock, K., Callaghan, D., Gorton, R., Kovacs, E. M., Lyons, M., Mumby, P., & Roelfsema, C. (2021). Baseline mapping to support reef restoration and resilience-based management in the Whitsundays. https://doi.org/10.13140/RG.2.2.26976.20482

Bozec, Y.-M., Hock, K., Mason, R. A. B., Baird, M. E., Castro-Sanguino, C., Condie, S. A., Puotinen, M., Thompson, A., & Mumby, P. J. (2022). Cumulative impacts across Australia's Great Barrier Reef: A mechanistic evaluation. Ecological Monographs, 92(1), e01494. https://doi.org/10.1002/ecm.1494